JPEG: Fix setjmp() issue (thanks to Sven Neumann)
Wrapping the call to setjmp() into a helper function is a bad
idea as setjmp() saves the stack for later use by longjmp().

The way it was done here totally broke error handling for fatal
errors from libjpeg. Reverting this part of commit aaebf60b86
restores error handling in the JPEG image provider.